Updates to the straw tracker configuration files.
Changing the width to be a constant instead of in a vector.
Taking out the information for the support plate (since it's now
gone in the new model). Taking out the piping information becaues it's
now gone. Adding in a value called strawStationPiping that will hold
the distance between the stations.