Changed the function Plane() to InnerRow() where Row runs 0-3.
Also created TotalPlane to give you the total Plane number where
Plane is now refered to as the view (two Planes per module, two layers
per plane). This is used in some gm2analyses. Previously the total plane
was refering to each individual row of straws. Now the plane is the group
of two (the view). Changes made accordingly and the functions in the
analysis now care about the Plane and not the row of wires.